requirements.txt
streamlit
langchain>=0.3.6
langchain-openai
langchain-community
PyPDF2
python-dotenv
faiss-cpu
